simplify db connection string
