+++ .env.local
@@ -9,3 +9,4 @@ INPUT_DEPLOYMENT_PATH='deploy/apply/live'
 INPUT_TEMP_RESOURCES_DIR = 'temp'
 INPUT_PR_NUMBER = '2222'
 INPUT_PR_REPOSITORY = 'nile-repo-temp'
+INPUT_CONNECTION_STRING='mongodb://admin:admin@localhost:27017/cicd?authSource=admin'
